codegen: relocate Polkadot, flatten encoding/events/statements/optimize; seal TargetCodegen
