+++ src/abi/polkadot.rs
@@ -22,8 +22,8 @@ use semver::Version;
 use solang_parser::pt;
 
 use crate::{
-    codegen::polkadot::SCRATCH_SIZE,
     codegen::revert::{SolidityError, ERROR_SELECTOR, PANIC_SELECTOR},
+    codegen::targets::polkadot::return_code::SCRATCH_SIZE,
     sema::{
         ast::{self, ArrayLength, EventDecl, Function},
         tags::render,
