--- src/codegen/dispatch/mod.rs
@@ -1,3 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-pub(crate) mod polkadot;
