+++ src/codegen/optimize/array_boundary.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use super::vartable::Vartable;
 use crate::codegen::cfg::{ControlFlowGraph, Instr};
+use crate::codegen::vartable::Vartable;
 use crate::codegen::Expression;
 use crate::sema::ast::Type;
 use solang_parser::pt::Loc;
