+++ src/codegen/optimize/dead_storage.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use super::cfg::{BasicBlock, ControlFlowGraph, Instr};
+use crate::codegen::cfg::{BasicBlock, ControlFlowGraph, Instr};
 use crate::codegen::Expression;
 use crate::sema::ast::{Namespace, RetrieveType, Type};
 use solang_parser::pt::Loc;
