+++ src/codegen/optimize/mod.rs
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod array_boundary;
+pub(crate) mod constant_folding;
+pub(crate) mod dead_storage;
+pub(crate) mod reaching_definitions;
+pub(crate) mod strength_reduce;
+pub(crate) mod subexpression_elimination;
+pub(crate) mod undefined_variable;
+pub(crate) mod unused_variable;
+pub(crate) mod vector_to_slice;
