+++ src/codegen/optimize/reaching_definitions.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use super::cfg::{BasicBlock, ControlFlowGraph, Instr};
+use crate::codegen::cfg::{BasicBlock, ControlFlowGraph, Instr};
 use crate::codegen::Expression;
 use indexmap::IndexMap;
 use std::collections::HashSet;
