+++ src/codegen/optimize/strength_reduce/value.rs
