+++ src/codegen/optimize/undefined_variable.rs
