+++ src/codegen/targets/polkadot/events.rs
@@ -4,8 +4,8 @@ use std::vec;
 
 use crate::codegen::cfg::{ControlFlowGraph, Instr};
 use crate::codegen::encoding::abi_encode;
-use crate::codegen::events::EventEmitter;
 use crate::codegen::expression::expression;
+use crate::codegen::interface::EventEmitter;
 use crate::codegen::interface::TargetCodegen;
 use crate::codegen::vartable::Vartable;
 use crate::codegen::{Builtin, Expression, Options};
