+++ src/codegen/targets/polkadot/return_code.rs
@@ -125,7 +125,7 @@ impl RetCodeCheck {
 ///
 /// If `bubble_up` is set to true, this will revert the contract execution on failure.
 /// Otherwise, the expression comparing the return code against `0` is returned.
-pub(super) fn check_transfer_ret(
+pub(crate) fn check_transfer_ret(
     loc: &Loc,
     success: usize,
     cfg: &mut ControlFlowGraph,
